requirements/req_load_new_data.txt
requests==2.31.0
feedparser==6.0.10
newspaper3k==0.2.8